requirements.txt

streamlit==1.40.0
python-dotenv==1.0.0
ffmpeg-python==0.2.0
pydub==0.25.1

openai==1.75.0
tqdm==4.67.1
jiter==0.9.0